rename meta/timeout module to meta/time
